Various fixes (#274)
* Fix ESLint warnings

ESLint was giving a [no-invalid-this](https://eslint.org/docs/rules/no-invalid-this) warning. The annotations.display callback isn't documented as having a `this` parameter, and I don't see any way that `this` could work in this context, so I removed it.

* Update plugin hooks for new IPlugin interface

https://www.chartjs.org/docs/master/getting-started/v3-migration#iplugin-interface

This fixes problems with `handleEvent` not seeing the correct event type and with the drawTime option not being honored.

* Fix moveListened vs. listened

If a chart's only hooks were move hooks on individual annotations, then when the chart is updated, `state.listened` was cleared at the beginning of `updateListeners`, but since `state.moveListened` was still set, the `annotations.forEach` loop didn't re-evaluate move listeners and re-set `state.listened`.