Type fixes
Allow `annotations` to be an object (as used in the current documentation) as well as an array (as used by older versions of the plugin).

Allow `id` property, as set by the object form of `annotations`.

Suppress ESLint warning; TypeScript declaration merging must exactly match the original, so ESLint's warning that the type is unused is a false positive.

Fix a typo in options.md.